Extract reifiers/has_many
