+++ lib/paper_trail/reifier.rb
@@ -1,5 +1,6 @@
 require "paper_trail/attribute_serializers/object_attribute"
 require "paper_trail/reifiers/belongs_to"
+require "paper_trail/reifiers/has_many"
 require "paper_trail/reifiers/has_one"
 
 module PaperTrail
@@ -125,7 +126,7 @@ module PaperTrail
         }
         versions = load_versions_for_hmt_association(assoc, ids, tx_id, options[:version_at])
         collection = Array.new assoc.klass.where(assoc.klass.primary_key => ids)
-        prepare_array_for_has_many(collection, options, versions)
+        Reifiers::HasMany.prepare_array(collection, options, versions)
         collection
       end
 
@@ -149,22 +150,6 @@ module PaperTrail
           first
       end
 
-      # Given a `has_many` association on `model`, return the version records
-      # from the point in time identified by `tx_id` or `version_at`.
-      # @api private
-      def load_versions_for_hm_association(assoc, model, version_table, tx_id, version_at)
-        version_id_subquery = ::PaperTrail::VersionAssociation.
-          joins(model.class.version_association_name).
-          select("MIN(version_id)").
-          where("foreign_key_name = ?", assoc.foreign_key).
-          where("foreign_key_id = ?", model.id).
-          where("#{version_table}.item_type = ?", assoc.class_name).
-          where("created_at >= ? OR transaction_id = ?", version_at, tx_id).
-          group("item_id").
-          to_sql
-        versions_by_id(model.class, version_id_subquery)
-      end
-
       # Given a `has_many(through:)` association and an array of `ids`, return
       # the version records from the point in time identified by `tx_id` or
       # `version_at`.
@@ -181,7 +166,7 @@ module PaperTrail
           ).
           group("item_id").
           to_sql
-        versions_by_id(assoc.klass, version_id_subquery)
+        Reifiers::HasMany.versions_by_id(assoc.klass, version_id_subquery)
       end
 
       # Reify onto `model` an attribute named `k` with value `v` from `version`.
@@ -213,59 +198,6 @@ module PaperTrail
         attrs.each do |k, v|
           reify_attribute(k, v, model, version)
         end
-      end
-
-      # Replaces each record in `array` with its reified version, if present
-      # in `versions`.
-      #
-      # @api private
-      # @param array - The collection to be modified.
-      # @param options
-      # @param versions - A `Hash` mapping IDs to `Version`s
-      # @return nil - Always returns `nil`
-      #
-      # Once modified by this method, `array` will be assigned to the
-      # AR association currently being reified.
-      #
-      def prepare_array_for_has_many(array, options, versions)
-        # Iterate each child to replace it with the previous value if there is
-        # a version after the timestamp.
-        array.map! do |record|
-          if (version = versions.delete(record.id)).nil?
-            record
-          elsif version.event == "create"
-            options[:mark_for_destruction] ? record.tap(&:mark_for_destruction) : nil
-          else
-            version.reify(
-              options.merge(
-                has_many: false,
-                has_one: false,
-                belongs_to: false,
-                has_and_belongs_to_many: false
-              )
-            )
-          end
-        end
-
-        # Reify the rest of the versions and add them to the collection, these
-        # versions are for those that have been removed from the live
-        # associations.
-        array.concat(
-          versions.values.map { |v|
-            v.reify(
-              options.merge(
-                has_many: false,
-                has_one: false,
-                belongs_to: false,
-                has_and_belongs_to_many: false
-              )
-            )
-          }
-        )
-
-        array.compact!
-
-        nil
       end
 
       # @api private
@@ -315,27 +247,12 @@ module PaperTrail
         reify_has_many_through_associations(transaction_id, assoc_has_many_through, model, options)
       end
 
-      # Reify a single, direct (not `through`) `has_many` association of `model`.
-      # @api private
-      def reify_has_many_association(assoc, model, options, transaction_id, version_table_name)
-        versions = load_versions_for_hm_association(
-          assoc,
-          model,
-          version_table_name,
-          transaction_id,
-          options[:version_at]
-        )
-        collection = Array.new model.send(assoc.name).reload # to avoid cache
-        prepare_array_for_has_many(collection, options, versions)
-        model.send(assoc.name).proxy_association.target = collection
-      end
-
       # Reify all direct (not `through`) `has_many` associations of `model`.
       # @api private
       def reify_has_many_associations(transaction_id, associations, model, options = {})
         version_table_name = model.class.paper_trail.version_class.table_name
         each_enabled_association(associations) do |assoc|
-          reify_has_many_association(assoc, model, options, transaction_id, version_table_name)
+          Reifiers::HasMany.reify(assoc, model, options, transaction_id, version_table_name)
         end
       end
 
@@ -424,22 +341,6 @@ module PaperTrail
         inher_col_value = attrs[inheritance_column_name]
         class_name = inher_col_value.blank? ? version.item_type : inher_col_value
         class_name.constantize
-      end
-
-      # Given a SQL fragment that identifies the IDs of version records,
-      # returns a `Hash` mapping those IDs to `Version`s.
-      #
-      # @api private
-      # @param klass - An ActiveRecord class.
-      # @param version_id_subquery - String. A SQL subquery that selects
-      #   the IDs of version records.
-      # @return A `Hash` mapping IDs to `Version`s
-      #
-      def versions_by_id(klass, version_id_subquery)
-        klass.
-          paper_trail.version_class.
-          where("id IN (#{version_id_subquery})").
-          inject({}) { |a, e| a.merge!(e.item_id => e) }
       end
     end
   end
